Rename clasname and fix namespaces
